+++ extra/scripts/phase-5.ps1
@@ -1,0 +1,4 @@
+# Post updates entries, not required to do a reboot
+Write-Output "Start Phase 5"
+Write-Output "End Phase 5"
+