+++ extra/scripts/phase-5b.docker.ps1
@@ -1,24 +1,33 @@
-# Post updates entries, not required to do a reboot
 # Enable docker
 $requiredVersion="18.09"
 $installCompose = $true
 
 Write-Output "Start Phase 5b-Docker"
 try {
-  Install-Module -Name DockerMsftProvider -Repository PSGallery -Force
-  Update-Module DockerMsftProvider -Force
+  Install-Module -Name DockerMsftProvider -Repository PSGallery -Force -Confirm:$false
+  Update-Module DockerMsftProvider -Force -Confirm:$false
+}
+catch {
+  Write "Phase 5b-docker failed - Install/Update module problem"
+  exit (-1)
+}
+
+try {
   Install-Package -Name docker -ProviderName DockerMsftProvider -RequiredVersion $RequiredVersion -Confirm:$false -Force
   Set-Service -Name docker -StartupType Automatic
 }
 catch {
-  Write "Phase 5b-docker failed"
+  Write "Phase 5b-docker failed - Install-Package Docker problem"
   exit (-1)
 }
+
+
 if ($installCompose) {
   try {
     choco install docker-compose -y
   }
   catch {
+    Write "Phase 5b-docker failed - Install docker-compose problem"
   }
 }
 Write-Output "End Phase 5b"