+++ extra/scripts/phase-5b.ps1
@@ -1,5 +1,0 @@
-# Post updates entries, not required to do a reboot
-
-# Spectre meltdown
-Write-Output "Start Phase 5b"
-Write-Output "End Phase 5b"